chore(attribute-fetchers): rename AttributeFetcher to AttributeFetchers
